Move over to http2-support for now, pull in impl from existing single connection.
